automation/requirements.txt
google-generativeai
requests
python-dotenv
google-cloud-aiplatform
